refactor: extra checker and parser
